Project initialization #4

+++ backend/src/main/java/com/project/leukbook/mapper/UserBodyModelMapper.java
@@ -16,5 +16,5 @@ public interface UserBodyModelMapper {
 
     int updateByPrimaryKey(UserBodyModel record);
 
-    UserBodyModel selectByUserId(Integer userId);
+    UserBodyModel selectByUserId(Long userId);
 }
